Truncate the manifest if there are too many object entries
Normally, there shouldn't be many object entries in the manifest since new
entries are added only if an include file has changed but not the source file,
and you typically change source files more often than header files. However,
it's certainly possible to imagine cases where the manifest will grow large
(for instance, a generated header file that changes for every build), and this
must be taken care of since processing an ever growing manifest eventually will
take too much time. A good way of solving this would be to maintain the object
entries in LRU order and discarding the old ones. An easy way is to throw away
all entries when there are too many. Let's do that for now.