Bail out on profiling flags
GCCs options -fprofile-generate and -fprofile-use are used for profile guided
optimization. It depends on the ability to locate the profile data files
(.gcda) after running the executable with the training data. However, ccache
prevents the compiler from finding the correct profile data file. Therefore the
following patch disables the caching when one of the -fprofile-generate
-fprofile-use flags is found.