Avoid preprocessed hits when using precompiled headers
Otherwise, we would get a false hit if the PCH has changed (but the rest of
the source has not) since the preprocessed output doesn't include the PCH
content.

The PCH content is now also included in the preprocessor output hash to
avoid several manifest entries (where the PCH hash differs) pointing to the
same object file.