add support for '@' parameters
These indicate to the compiler that additional command line options
should be read from a text file. If encountered, read the file,
tokenize any arguments, and if any are found, do an in-place replacement
of the '@' parameter with the arguments within the file.

args_insert() added to insert a set of arguments into a position within
another set of arguments.

args_init_from_gcc_atfile() reads and processes the argument files using the
quoting/escape conventions that GCC expects.

Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>