Only clean dist generate targets when using distclean
Some file are only generated when making a distribute, such as docs, and should
only be cleaned when using the distclean target. If we clean the docs with the
normal clean target then when using the Makefile with the distributed source
code will lead to an error if you configure, make, make clean, and then make
again.