Fix bug testing with ecryptfs
The test script assumes that zero-length files take up zero disk space,
that the block size is 4k, and that the disk usage will be the file size
rounded up to the nearest block size. Unfortunately these were not true for
ecryptfs.

Additionally, running the test.sh from another directory did not work due
to a bad location assumption in another test.