Speed up copy4 and copy64 on little-endian systems
The copy64 function implements an endian-safe copy routine for an array of
16 32-bit integers, but this is sub-optimal on machines where the
byte-order is already correct. Likewise for copy4.

This patch replaces the implementation with a simple memcpy when
appropriate, and retains the old implementation otherwise.

Note that the compiler will always inline calls to memcpy for small
byte-counts, so this is a big win.