Remove incorrect preconditions for outputting captured stderr
errno is set in an unrelated system call. In my tests involving compiler
failures, it is last set to ENOENT in from_cache() when calling stat() for the
cached object, accidentally making "errno == ENOENT" true.

The condition was no longer necessary as of 18a645451 ("Create destination file
and then copy into cache instead of the opposite"), as the temp output file no
longer needed to be moved.