Don't delete the clang diagnostics file after compiler failures
move_file() returns an error when called with the same source and destination,
causing the diagnostics file to be removed.

As of 18a645451 ("Create destination file and then copy into cache instead of
the opposite") it is no longer necessary to move the diagnostics file at all
after a compiler failure.