Fix regression in recent change related to zlib in nonstandard location
To allow for specifying a -L flag in LDFLAGS and have it take effect for
-lz, 0220de9c8ebfb18caae2ac1aa163d060e98ceade put -lz in extra_libs
instead of extra_ldflags. However, extra_libs is supposed to contain
paths to libraries which are prerequisites of the ccache link rule, and
some older versions of make got confused by seeing -lz as a
prerequisite.

This fix is to revert 0220de9c8ebfb18caae2ac1aa163d060e98ceade and
instead make sure that LDFLAGS comes before extra_ldflags.