Count the number of cleanups performed, in stats
When running a busy ccache, it is useful to keep track of the number
of cleanups performed to know when you need to increase the cache size.

The number shown is actually the number of subdirectories, so when
you do a full cleanup the counter usually increases by 16 (not just 1).