memcached: Fix bug when getting object from memcached in from_fscache
If memcached_only is false and the object is not in the fs cache, we try to
fetch it from memcached. However, the "invalid (empty) object file" check
will then trigger and remove the object from the fs cache again since
the object size is read from a (potentially) uninitialized struct stat.