Removed time_of_compilation check wrt SLOPPY_FILE_STAT_MATCHES
The whole code seems to be a thinko.  For a hit, neither ctime
nor mtime should be greater than or equal to time_of_compilation.
The code only seems to work because time_of_compilation is 0
at this stage (i.e. has not been initialized).

While at it, I also introduce an optimization: when sizes do
not match, it's a good chance to bail out early; there is no
point in further hashing the file.