Treat unreadable conf file like missing instead of a fatal error
This reverts 0b18af47 and implements a better solution.

The major reason is to keep ccache transparent, i.e. to still have the
following behavior:

* If ccache returns exit code == 0, then any produced stderr comes from
  the compiler and only from the compiler.
* If ccache returns exit code != 0, then ccache may print error messages
  of its own.

The reason is that autoconf configure scripts have been known to fail
tests if the compiler emits anything to stderr even if the produced
result actually works. (And printing to stdout is also a no-no.)