Add small helper to run all Travis tests
The idea behind this script is to run the same checks as .travis.yml
but locally instead. It is not identical, but should be close enough.

Note: this does not include any tests for the osx or cuda platforms,
only the linux targets - although that does include the mingw build.