Put "calculate preprocessor hash" code into "if" block
In a following commit we will execute this code conditionally.

Even though the change is simple, the diff looks rather ugly (unless
when ignoring whitespace differences).  Having this step as a separate
commit should make the following commit more clear.

Chose to not factor it out into a separate function because of the
amount of local variables referenced.