Print compiler's stderr when compiling with memcached_only=true
In the memcached_only mode no files are created in the local cache. All
compilation results are stored in memcached backend but the routine
responsible for running a compiler and storing its outputs -
to_memcached - still used send_cached_stderr to present compiler's
stderr to the user. The latter function relied on the existence of
cached stderr in the local cache printing nothing.

The fix is to print the stderr data sent to memcached backend instead.