Minor fixes for cygwin/msys2
Add -Wno-error=implicit-fallthrough, if the compiler supports it,  to
dev.mk so that zlib will compile with -Werror.

Disable -Wdeprecated-declarations in the mkstemp() shim in util.c so
that it compiles with -Werror.

Stop assuming that mkstemp() is broken on cygwin, I could not find any
documentation for this.

Fix a couple of erroneous checks in zlib that assume cygwin is the same
as win32.