Two changes are contained in this commit:
1) Depend mode fix for a case with unchaged common hash (unchanged source file and other parameters
that affect the hash), but with some changes in the header file[s] the source file depends on.
This allows for multiple versions of the object file and dependency file to exist in the cache for a single source file.
Added simple unit test in test/suites/depend_different_headers.bash to verify this behavior.

2) Depend mode also forces the paths of the include files to be relative if the base dir is specified. This
was previously only done if preprocessor was run.