Depend mode - support multiple versions of dependencies
Depend mode fix for a case with unchaged common hash (unchanged source file and other parameters
that affect the hash), but with changes in the header file[s] the source file depends on.
This allows for multiple versions of the object file and dependency file to exist for one common hash.
Added unit tests in test/suites/depend_different_headers.bash to test this behavior.