do not resolve symlinks for CCACHE_BASEDIR
When building with builddir != srcdir, it is practically necessary
to have either builddir contain srcdir or the other way around,
otherwise relative paths from one to another will not be the same
for different builds, which will lead to misses with CCACHE_BASEDIR.
In case one wants builddir and srcdir to be elsewhere (for a number
of reasons, different filesystems, build files inside srcdir being
a mess, ...), this setup can be achieved using a symlink. But then
symlink resolving changes the relative path to exactly the broken
setup mentioned above.

The resolving was introduced in 2df269a3121889ebcdfa5d98dfb4d675f690e039
with questionable reasoning (how can CWD possibly contain "/./"?).
The only test that breaks by removing this change is an artifical
test designed to test exactly this (mis)feature, all the rest work fine.