Make new installations use XDG base directory
This effectively deprecates ~/.ccache as the default directory.

It is still possible to revert to the old behaviour by either setting
this directory in the secondary config file, or by setting the
CCACHE_DIR environment variable.

The new directories are ~/.config/ccache for the configuration file, and
~/.cache/ccache for all other files.  We may want to introduce a third
directory, ~/.local/share/ccache, for statistics, but I wanted to
validate the new behaviour first.

Fixes #191.